Add/update/cleanup copyright headers in scimath (#116)
* DEV : Add/update/cleanup copyright headers in scimath

This commits adds/updates/cleansup copyright headers in scimath.
A large number of modules didnt have copyright headers. Some modules had
copyright headers which were outdated or which had copyright headers in
an old format. A smaller number had the original copyright of the author
whose work was derived in some modules.

* CLN : Remove hashbang lines at the top of files

these are unnecessary and these were interfering when running flake8_ets
on the codebase

	modified:   scimath/units/SI.py
	modified:   scimath/units/area.py
	modified:   scimath/units/energy.py
	modified:   scimath/units/force.py
	modified:   scimath/units/power.py
	modified:   scimath/units/pressure.py
	modified:   scimath/units/substance.py
	modified:   scimath/units/temperature.py
	modified:   scimath/units/unit.py
	modified:   scimath/units/unit_parser.py