Don't test on macOS/EDM (#177)
macOS NumPy eggs for EDM are currently incompatible with GitHub-hosted
macOS runners: the EDM eggs use AVX2 instructions, which the runners
don't support. As a result, our CI jobs are currently failing.

This PR simply turns off the test runs on macOS/EDM. Note that we're
still testing on macOS in the `test-with-pypi.yml` workflow.

Addendum: also updates versions of the various actions used in the
workflow so that they'll keep on working later this year.